+++ CHANGELOG.md
@@ -327,6 +327,11 @@ confluent-kafka-python is based on librdkafka v2.2.0, see the
 for a complete list of changes, enhancements, fixes and upgrade considerations.
 
 
+## Next Version
+
+ - Added `consumer.io_event_enable` to integrate with event-loops like asyncio.
+   (#1448)
+
 ## v2.1.1
 
 v2.1.1 is a maintenance release with the following fixes and enhancements: